New feature: Export of profile data to workspace
The profile("2d")-associated data along one user-selected dimension can
now be exported to the workspace (similar to the ROI-data export). The
variable with be 3D (profile width and length plus extra dimension) or
4D in case of an RGB profile. This makes it easy to visualize e.g. how a
profile changes over time (if the user-selected dimension is time). So
far only the data associated with the first data set is exported.
Changeing this should be very easy.